Cybersecurity Fundamentals
In this folder you will find the start of my cybersecurity career with our class CIT 285, Cybersecurity Fundamentals. In this class we covered an introduction to fundamental concepts and technologies in cybersecurity, with lab exercises that focused on practical aspects of securing computers. The topics include were risk analysis, common attacks and defenses, authentication, access control, network security, application and data security, and cryptography.