Advanced Cybersecurity Skills
In this folder you will find my advanced cybersecurity skillset that was worked on in our CIT 485 class, Advanced Cybersecurity. This course was provided to help develop baseline skills needed to proactively capture, monitor, and respond to network traffic findings and emphasize software and application security, automation, and threat hunting, and IT regulatory compliance, which affects the daily work of security analysts. This course's primary focus
was to provide targeted content and integrated hands-on skills to apply behavioral analytics to networks and
devices to prevent, detect, and combat cybersecurity threats through continuous security monitoring. This course was used to prepare students for the CompTIA CySA+ certification exam and also used the CompTIA CySA+ book as a guide and reference for the course.